metrics is moved under executor
