+++ clients/mqtt/mqtt.go
@@ -8,7 +8,7 @@ import (
 	paho "github.com/eclipse/paho.mqtt.golang"
 	"github.com/gobench-io/gobench/dis"
 	"github.com/gobench-io/gobench/executor"
-	"github.com/gobench-io/gobench/metrics"
+	"github.com/gobench-io/gobench/executor/metrics"
 )
 
 const conTotal string = "mqtt.connection.current_total"
